+++ src/main/java/ru/stockmann/replenishment/services/cddata/process/CDDataStageRow.java
@@ -35,7 +35,7 @@ public record CDDataStageRow(
         BigDecimal salesDiscount,
         BigDecimal stockStoresPcs,
         BigDecimal stockStoresDd,
-        Integer planRub,
+        BigDecimal planRub,
         String draiveryCd,
         String skuColorRus,
         String skuComposition,
